requirements.txt
aiohttp[speedups]
python-socketio
python-socketio[client]
opencv-python
pyinstaller
